카드뷰 객체 파일로 분리
